refactor: PointCalculator 로직 ChessBoard 내 병합
