+++ src/main/java/chess/model/board/ChessBoard.java
@@ -1,18 +1,27 @@
 package chess.model.board;
 
+import static java.util.stream.Collectors.collectingAndThen;
+import static java.util.stream.Collectors.toMap;
+
 import chess.dto.ChessBoardDto;
 import chess.dto.PieceDto;
 import chess.model.piece.Empty;
 import chess.model.piece.Piece;
 import chess.model.piece.Side;
 import chess.model.position.ChessPosition;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 public class ChessBoard {
     private static final int NORMAL_KING_COUNT = 2;
+    private static final int PAWN_COUNT_WHEN_SOLID_IN_FILE = 1;
+    private static final double POINT_WHEN_SAME_FILE_PAWN = -0.5;
 
     private final Map<ChessPosition, Piece> board;
     private final Side turn;
@@ -53,15 +62,41 @@ public class ChessBoard {
         return findPiece(position).isEnemy(side);
     }
 
-    public PointCalculator getPointCalculator() {
-        return new PointCalculator(Collections.unmodifiableMap(board));
-    }
-
     public ChessBoardDto convertDto() {
         return ChessBoardDto.from(board.entrySet()
                 .stream()
                 .map(entry -> PieceDto.from(entry.getKey(), entry.getValue()))
                 .collect(Collectors.toSet()), turn.name());
+    }
+
+    public Points calculate() {
+        return Arrays.stream(Side.values())
+                .filter(side -> !side.isEmpty())
+                .collect(collectingAndThen(toMap(Function.identity(), this::calculatePiecePoints), Points::new));
+    }
+
+    private boolean hasSameFilePawn(final ChessPosition position, final Set<ChessPosition> positions) {
+        return calculatePawnCountSameFile(position, positions) != PAWN_COUNT_WHEN_SOLID_IN_FILE;
+    }
+
+    private int calculateKingCount() {
+        return (int) board.values()
+                .stream()
+                .filter(Piece::isKing)
+                .count();
+    }
+
+    private long calculatePawnCountSameFile(final ChessPosition position, final Set<ChessPosition> positions) {
+        return positions.stream()
+                .filter(position::hasSameFile)
+                .count();
+    }
+
+    private long getSameFilePawnCount(final Side side) {
+        final Set<ChessPosition> positions = findAllSamSidePawns(side);
+        return positions.stream()
+                .filter(position -> hasSameFilePawn(position, positions))
+                .count();
     }
 
     private void validatePieces(final Piece sourcePiece, final Piece targetPiece) {
@@ -95,15 +130,36 @@ public class ChessBoard {
         board.put(sourcePosition, new Empty());
     }
 
-    private int calculateKingCount() {
-        return (int) board.values()
-                .stream()
-                .filter(Piece::isKing)
-                .count();
-    }
-
     private Piece findPiece(final ChessPosition position) {
         return board.getOrDefault(position, new Empty());
+    }
+
+    private Point calculatePiecePoints(final Side side) {
+        final Point totalSum = calculateTotalSum(side);
+        final Point minusPoint = calculateSameFilePawn(side);
+        return totalSum.sum(minusPoint);
+    }
+
+    private Point calculateTotalSum(final Side side) {
+        return board.values()
+                .stream()
+                .filter(piece -> piece.isSameSide(side))
+                .map(Piece::getPoint)
+                .reduce(Point.getDefaults(), Point::sum);
+    }
+
+    private Point calculateSameFilePawn(final Side side) {
+        final long sameFilePawn = getSameFilePawnCount(side);
+        return Point.from(sameFilePawn * POINT_WHEN_SAME_FILE_PAWN);
+    }
+
+    private Set<ChessPosition> findAllSamSidePawns(final Side side) {
+        return board.entrySet()
+                .stream()
+                .filter(entry -> entry.getValue().isSameSide(side))
+                .filter(entry -> entry.getValue().isPawn())
+                .map(Entry::getKey)
+                .collect(Collectors.toSet());
     }
 
     public Map<ChessPosition, Piece> getBoard() {
